% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan.normalize}
\alias{vulcan.normalize}
\title{Function to normalize promoter peak data}
\usage{
vulcan.normalize(vobj)
}
\arguments{
\item{vobj}{a list, the output of the \code{'vulcan.annotate'} function}
}
\description{
This function normalizes gene-centered ChIP-Seq data using VST
}
