% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan.annotate}
\alias{vulcan.annotate}
\title{Function to annotate peaks for VULCAN analysis}
\usage{
vulcan.annotate(vobj, lborder = -10000, rborder = 10000,
  method = c("closest", "strongest", "sum", "topvar", "farthest", "lowvar"))
}
\arguments{
\item{vobj}{A list of peakcounts, samples and peakrpkms (i.e. the output of
the funcion vulcan.import)}

\item{lborder}{Boundary for peak annotation (in nucleotides) upstream of
the Transcription starting site (default: -10000)}

\item{rborder}{Boundary for peak annotation (in nucleotides) downstream of
the Transcription starting site (default: 10000)}

\item{method}{Method to deal with multiple peaks found within gene promoter
boundaries.
One of "closest","strongest","sum"(default),"topvar","farthest","lowvar"}
}
\value{
a list
}
\description{
This function coalesces and annotates a set of BAM files into peak-centered
data
}
\examples{
require(vulcandata)
vfile<-"deleteme.csv"
vulcandata::vulcansheet(vfile)
vobj<-vulcan.import(vfile)
unlink(vfile)
vobj<-vulcan.annotate(vobj,lborder=-10000,rborder=10000,method="sum")
}
