% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan.import}
\alias{vulcan.import}
\title{Function to import BAM files}
\usage{
vulcan.import(sheetfile, intervals = NULL)
}
\arguments{
\item{sheetfile}{path to a csv annotation file containing sample information
and BAM location}

\item{intervals}{size of the peaks. If NULL (default) it is inferred from
the average fragment length observed in the dataset}
}
\description{
This function coalesces and annotates a set of BAM files into peak-centered
data
}
