% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan.normalize}
\alias{vulcan.normalize}
\title{Function to normalize promoter peak data}
\usage{
vulcan.normalize(vobj)
}
\arguments{
\item{vobj}{a list, the output of the \code{"vulcan.annotate"} function}
}
\value{
a list
}
\description{
This function normalizes gene-centered ChIP-Seq data using VST
}
\examples{
library(vulcandata)
vfile<-"deleteme.csv"
vulcandata::vulcansheet(vfile)
vobj<-vulcan.import(vfile)
unlink(vfile)
vobj<-vulcan.annotate(vobj,lborder=-10000,rborder=10000,method="sum")
vobj<-vulcan.normalize(vobj)
}
