% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan}
\alias{vulcan}
\title{VULCAN - VirtUaL Chipseq data Analysis using Networks}
\usage{
vulcan(vobj, network, contrast, annotation = NULL, minsize = 10)
}
\arguments{
\item{vobj}{a list, the output of the \code{"vulcan.normalize"} function}

\item{network}{an object of class \code{"viper::regulon"}}

\item{contrast}{a vector of two fields, containing the condition names
to be compared (1 vs 2)}

\item{annotation}{an optional named vector to convert gene identifiers
(e.g. entrez ids to gene symbols)
Default (NULL) won't convert gene names.}
}
\value{
a list
}
\description{
This function calculates the enrichment of a gene regulatory network over a
ChIP-Seq derived signature
}
\examples{
require(vulcandata)
# Generate an annotation file from the dummy ChIP-Seq dataset
vfile<-"deleteme.csv"
vulcansheet(vfile)
# Import BAM and BED information into a list object
vobj<-vulcan.import(vfile)
unlink(vfile)
# Annotate peaks to gene names
vobj<-vulcan.annotate(vobj,lborder=-10000,rborder=10000,method="sum")
# Normalize data for VULCAN analysis
vobj<-vulcan.normalize(vobj)
# Detect which conditions are present
names(vobj$samples)

# Load an ARACNe network
# This is a regulon object as specified in the VIPER package, named "network"
load(system.file("extdata","network.rda",package="vulcandata",mustWork=TRUE))
# Run VULCAN
# We can reduce the minimum regulon size, since in this example only one
# chromosome
# was measured, and the networks would otherwise have too few hits
vobj_analysis<-vulcan(vobj,network=network,contrast=c("t90","t0"),minsize=5)
# Visualize output using the msviper plotting function
plot(vobj_analysis$msviper,mrs=7)

}
